Read wide numbers whole on 32-bit platforms
On 386 the suite failed in three places, all numbers wider than a 32-bit
int:

- A metazone use that has not ended runs, in ICU's data, to 9999-12-31:
  4,223,371,680 minutes, which wrapped negative. No winter date in London
  or Algiers then fell in its metazone, so "Greenwich Mean Time" was
  written "GMT+00:00", in some 5,800 cases of the sweeps.
- Collation's reordering bounds are 32-bit primaries, 0xFE000000, read
  through int.
- The Hebrew calendar counted the parts of a day since its epoch, some
  5e10, in int. ICU counts the year's start in int64 and estimates the
  year in doubles; go-intl now does both as ICU does.

blob.Reader.Uint now fails on a number an int cannot hold instead of
wrapping it, which is how the collation case was found, and Uint64 reads
the tables that hold such numbers. The encoding is unchanged, so no data
is regenerated.

The whole suite passes with GOARCH=386 as well as amd64.

+++ internal/zonegen/meta.go
@@ -174,12 +174,12 @@ func readMeta(archive *zip.ReadCloser) (*zonedata.Meta, error) {
 
 // minutes turns metaZones.txt's "1970-01-01 00:00", which is UTC, into
 // minutes since 1970 plus one, so that zero can mean unbounded.
-func minutes(s string) (int, error) {
+func minutes(s string) (int64, error) {
 	t, err := time.Parse("2006-01-02 15:04", s)
 	if err != nil {
 		return 0, fmt.Errorf("metaZones.txt: %w", err)
 	}
-	m := int(t.Unix()/60) + 1
+	m := t.Unix()/60 + 1
 	if m < 1 {
 		return 0, fmt.Errorf("metaZones.txt: %s is before 1970", s)
 	}
